+++ src/strategies/community-strategy/v1/cult-strategy/example.json
