+++ src/strategies/community-strategy/v1/erc20-balance-of/example.json
