+++ src/strategies/community-strategy/v1/premia-strategy/example.js
