+++ src/strategies/community-strategy/v1/rep3-platform/membership-bulk-update/index.ts
@@ -1,7 +1,7 @@
-import ActionCallerV1 from '../../../actions/v1';
-import { ActionOnType } from '../../../actions/utils/type';
-import { StrategyParamsType } from '../../../types';
-import { getAllMembershipNfts } from '../../../utils/rep3';
+import ActionCallerV1 from '../../../../../actions/v1';
+import { ActionOnType } from '../../../../../actions/utils/type';
+import { StrategyParamsType } from '../../../../../types';
+import { getAllMembershipNfts } from '../../../../../utils/rep3';
 
 //TODO: Membership differentiation from expiry
