+++ src/strategies/community-strategy/v1/rep3-platform/rep3-is-member/example.json
