+++ src/strategies/community-strategy/v2/across-strategy/example.js
